+++ src/Traits/MenuBuilder.php
@@ -2,6 +2,8 @@
 
 namespace Speso\Ussd\Traits;
 
+use Illuminate\Support\Facades\Lang;
+
 trait MenuBuilder
 {
     public function text(string $text): static
@@ -14,6 +16,27 @@ trait MenuBuilder
     public function line(string $text): static
     {
         $this->content .= $text . PHP_EOL;
+
+        return $this;
+    }
+
+    public function trans(string $key, array $replace = [], ?string $locale = null): static
+    {
+        $this->content .= Lang::get($key, $replace, $locale);
+
+        return $this;
+    }
+
+    public function transLine(string $key, array $replace = [], ?string $locale = null): static
+    {
+        $this->content .= Lang::get($key, $replace, $locale) . PHP_EOL;
+
+        return $this;
+    }
+
+    public function transChoice(string $key, int $number, array $replace = [], ?string $locale = null): static
+    {
+        $this->content .= Lang::choice($key, $number, $replace, $locale);
 
         return $this;
     }